feat(commentCardFunctions): add addCommentButtonListeners, addPageNumListeners, and removePageNumListeners functions
